fix: extract a common unload sentry
