+++ src/common/router.js
@@ -1,3 +1,6 @@
+import { showConfirmation } from '#/common/ui';
+import { i18n } from './util';
+
 function parse(hash) {
   const [pathname, search = ''] = hash.split('?');
   const query = search.split('&').reduce((res, seq) => {
@@ -19,21 +22,22 @@ export const lastRoute = () => stack[stack.length - 1] || {};
 
 updateRoute();
 
-function updateRoute() {
+function updateRoute(noConfirm) {
   const hash = window.location.hash.slice(1);
-  if (!route.pinned) {
+  if (noConfirm || !route.confirmChange) {
     Object.assign(route, parse(hash));
   } else if (route.hash !== hash) {
     // restore the pinned route
-    setRoute(route.hash);
+    setRoute(route.hash, false, true);
+    route.confirmChange(hash);
   }
 }
 
 // popstate should be the first to ensure hashchange listeners see the correct lastRoute
 window.addEventListener('popstate', () => stack.pop());
-window.addEventListener('hashchange', updateRoute, false);
+window.addEventListener('hashchange', () => updateRoute(), false);
 
-export function setRoute(hash, replace) {
+export function setRoute(hash, replace, noConfirm) {
   let hashString = `${hash}`;
   if (hashString[0] !== '#') hashString = `#${hashString}`;
   if (replace) {
@@ -42,5 +46,30 @@ export function setRoute(hash, replace) {
     stack.push(Object.assign({}, route));
     window.history.pushState('', null, hashString);
   }
-  updateRoute();
+  updateRoute(noConfirm);
+}
+
+export function getUnloadSentry(onConfirm, onCancel) {
+  async function confirmPopState(hash) {
+    try {
+      // popstate cannot be prevented so we pin current `route` and display a confirmation
+      await showConfirmation(i18n('confirmNotSaved'));
+      setRoute(hash, false, true);
+      onConfirm?.();
+    } catch {
+      onCancel?.();
+    }
+  }
+  function toggle(state) {
+    const onOff = `${state ? 'add' : 'remove'}EventListener`;
+    global[onOff]('beforeunload', onUnload);
+    route.confirmChange = state && confirmPopState;
+  }
+  return toggle;
+}
+
+function onUnload(e) {
+  e.preventDefault();
+  // modern browser show their own message text
+  e.returnValue = i18n('confirmNotSaved');
 }
