+++ src/options/views/tab-settings/vm-css.vue
@@ -7,7 +7,7 @@
 </template>
 
 <script>
-import { showMessage } from '#/options/utils';
+import { showMessage } from '#/common/ui';
 import SettingText from '#/common/ui/setting-text';
 
 export default {
