+++ src/options/views/tab-settings/vm-editor.vue
@@ -11,7 +11,7 @@
 
 <script>
 import options from '#/common/options';
-import { showMessage } from '#/options/utils';
+import { showMessage } from '#/common/ui';
 import SettingText from '#/common/ui/setting-text';
 
 export default {
