+++ src/options/views/tab-settings/vm-template.vue
@@ -6,7 +6,7 @@
 </template>
 
 <script>
-import { showMessage } from '#/options/utils';
+import { showMessage } from '#/common/ui';
 import SettingText from '#/common/ui/setting-text';
 
 export default {
